% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert2dummy.R
\name{likert2dummy}
\alias{likert2dummy}
\title{Convert variable with 5-pt Likert-scale to dummy variable}
\usage{
likert2dummy(df, x)
}
\value{
A vector of numbers resulting from converting 5-pt scale to a dummy variable.
}
\description{
This function converts 5-pt Likert scales to dummy variables using the rule that a 4 or 5 on a Likert scale will = 1 on the dummy scale, and a 3, 2, or 1 on Likert scale will = 0 on a dummy scale.
}
\examples{
likert2dummy(gapminder, lifeExp)

}
